+++ MyApplication/app/src/main/AndroidManifest.xml
@@ -13,22 +13,22 @@
         android:theme="@style/Theme.MyApplication"
         tools:targetApi="31">
         <activity
-            android:name=".ViewRecycler"
+            android:name=".presentation.main.page1.ViewRecycler"
             android:exported="false"
             android:label="@string/title_activity_view_recycler"
             android:theme="@style/Theme.MyApplication" />
         <activity
-            android:name=".RegisterActivity"
+            android:name=".presentation.signup.RegisterActivity"
             android:exported="false"
             android:label="@string/title_activity_register"
             android:theme="@style/Theme.MyApplication" />
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.login.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MyApplication"></activity>
         <activity
-            android:name=".LoginActivity"
+            android:name=".presentation.main.LoginActivity"
             android:exported="true"
             android:label="login"
             android:theme="@style/Theme.MyApplication">
