+++ MyApplication/app/src/main/java/com/example/myapplication/presentation/main/LoginActivity.kt
@@ -1,44 +1,34 @@
-package com.example.myapplication
+package com.example.myapplication.presentation.main
 
-import android.content.Intent
 import android.os.Bundle
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
+import com.example.myapplication.R
 import com.example.myapplication.databinding.ActivityLoginBinding
+import com.example.myapplication.presentation.main.page1.Result1Fragment
+import com.example.myapplication.presentation.main.page2.Result2Fragment
 
 
 class LoginActivity : AppCompatActivity() {
     lateinit var binding: ActivityLoginBinding
-    lateinit var resultLauncher: ActivityResultLauncher<Intent>
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityLoginBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        //버튼 누르면 리사이클러뷰가 있는 xml로 이동
-        resultLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult())
-        {result ->
-            if (result.resultCode == RESULT_OK){}}
-        binding.btLoginRecyclerview.setOnClickListener {
-            val intent = Intent(this, ViewRecycler::class.java)
-            resultLauncher.launch(intent)
-        }
-
         binding.bnBottomNavigation.setOnItemSelectedListener { item ->
             when (item.itemId) {
                 R.id.item_fragment1 -> {
 
                     supportFragmentManager.beginTransaction()
-                        .replace(R.id.fg_result, Result1Fragment())
-                        .commit()
+                        .replace(R.id.fg_result, Result1Fragment())   //replace: 프레그먼트 교체
+                        .commit()  //commit(): 현재 상태 적용
                 }
 
                 R.id.item_fragment2 -> {
                     supportFragmentManager.beginTransaction()
                         .replace(R.id.fg_result, Result2Fragment())
-                        .commit()
+                        .commit()   
                 }
 
                 R.id.item_fragment3 -> {
