+++ MyApplication/app/src/main/java/com/example/myapplication/presentation/main/page1/UserData.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.presentation.main.page1
 
 data class UserData(
     val time: String,
