+++ MyApplication/app/src/main/java/com/example/myapplication/presentation/signup/RegisterActivity.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.presentation.signup
 
 import android.app.Activity
 import android.content.Intent
