+++ MyApplication/app/src/main/res/layout/activity_login.xml
@@ -22,11 +22,4 @@
         tools:layout_editor_absoluteX="0dp"
         tools:layout_editor_absoluteY="67dp" />
 
-    <Button
-        android:id="@+id/bt_login_recyclerview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="리사이클러뷰"
-        app:layout_constraintBottom_toTopOf="@id/bn_bottomNavigation"
-        app:layout_constraintStart_toStartOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
