+++ MyApplication/app/src/main/res/layout/fragment_result1.xml
@@ -1,15 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".Result1Fragment">
+    tools:context=".presentation.main.page1.Result1Fragment">
 
-    <TextView
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rv_recycler1"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="프래그먼트1 성공"
-        android:textSize="100sp" />
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+        tools:listitem="@layout/item_recycler_view"/>
 
-</FrameLayout>
+</LinearLayout>
