+++ MyApplication/app/src/main/res/layout/fragment_result2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Result2Fragment">
+    tools:context=".presentation.main.page2.Result2Fragment">
 
     <TextView
         android:layout_width="match_parent"
